Cleanup: removed containers/ConfigurationPage.js
